learning_agent/pyproject.toml
[project]
name = "crawl4ai-mcp"
version = "0.1.0"
description = "MCP server for integrating web crawling and RAG into AI agents and AI coding assistants"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai==0.6.2",
    "mcp>=1.10.0",
    "fastmcp>=2.0.0",
    "supabase==2.15.1",
    "openai==1.71.0",
    "python-dotenv",
    "sentence-transformers>=4.1.0",
    "neo4j>=5.28.1",
    "asyncpg>=0.29.0",
    "fastapi",
    "uvicorn",
    "redis>=5.0.0"
]

[tool.setuptools.packages.find]
include = ["aws*", "config*", "crawl*", "generators*", "models*", "routes*", "services*"]
